Updating for conversion to double precision.

+++ src/com/monyetmabuk/rajawali/tutorials/examples/general/TerrainFragment.java
@@ -31,7 +31,7 @@ public class TerrainFragment extends AExampleFragment {
 	private final class TerrainRenderer extends AExampleRenderer {
 		
 		private SquareTerrain mTerrain;
-		private float mLastY = 0f;
+		private double mLastY = 0;
 
 		public TerrainRenderer(Context context) {
 			super(context);
@@ -153,11 +153,11 @@ public class TerrainFragment extends AExampleFragment {
 			float distanceFromGround = 20;
 
 			for (int i = 0; i < 360; i += degreeStep) {
-				float radians = MathUtil.degreesToRadians(i);
-				float x = (float) Math.cos(radians) * (float) Math.sin(radians)
+				double radians = MathUtil.degreesToRadians(i);
+				double x = Math.cos(radians) * Math.sin(radians)
 						* radius;
-				float z = (float) Math.sin(radians) * radius;
-				float y = mTerrain.getAltitude(x, z) + distanceFromGround;
+				double z = Math.sin(radians) * radius;
+				double y = mTerrain.getAltitude(x, z) + distanceFromGround;
 
 				if (i > 0)
 					y = (y + mLastY) * .5f;
